requirements.txt
langchain==0.2.16 
langchain-core==0.2.38 
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-pinecone
pinecone
python-dotenv
fastapi
uvicorn
